move outer loop config to new package + fix raw use of OuterLoopConfig + use right static access for names
