+++ src/main/java/com/powsybl/openloadflow/OpenLoadFlowParameters.java
@@ -31,6 +31,12 @@ import com.powsybl.openloadflow.dc.equations.DcApproximationType;
 import com.powsybl.openloadflow.dc.equations.DcEquationSystemCreationParameters;
 import com.powsybl.openloadflow.graph.GraphConnectivityFactory;
 import com.powsybl.openloadflow.lf.AbstractLoadFlowParameters;
+import com.powsybl.openloadflow.lf.outerloop.config.AcOuterLoopConfig;
+import com.powsybl.openloadflow.lf.outerloop.config.DcOuterLoopConfig;
+import com.powsybl.openloadflow.lf.outerloop.config.DefaultAcOuterLoopConfig;
+import com.powsybl.openloadflow.lf.outerloop.config.DefaultDcOuterLoopConfig;
+import com.powsybl.openloadflow.lf.outerloop.config.ExplicitAcOuterLoopConfig;
+import com.powsybl.openloadflow.lf.outerloop.config.ExplicitDcOuterLoopConfig;
 import com.powsybl.openloadflow.network.*;
 import com.powsybl.openloadflow.network.util.PreviousValueVoltageInitializer;
 import com.powsybl.openloadflow.network.util.UniformValueVoltageInitializer;
