+++ src/main/java/com/powsybl/openloadflow/lf/outerloop/config/DcOuterLoopConfig.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.openloadflow;
+package com.powsybl.openloadflow.lf.outerloop.config;
 
 import com.powsybl.openloadflow.dc.DcOuterLoop;
